refactor: Llm component to use custom audio recorder hook
