+++ projects/portfolio/src/client/pages/Llm.tsx
@@ -1,81 +1,14 @@
-import { useEffect, useRef, useState } from 'react'
+import { useEffect, useState } from 'react'
+import { useAudioRecorder } from '../hooks/useAudioRecorder'
 
 export function Llm() {
-  const [isRecording, setIsRecording] = useState(false)
-  const [audioURL, setAudioURL] = useState<string | null>(null)
-  const [elapsedTime, setElapsedTime] = useState(0)
-  const [audioBlob, setAudioBlob] = useState<Blob | null>(null)
+  const { isRecording, audioURL, elapsedTime, audioBlob, handleToggleRecording } =
+    useAudioRecorder()
   const [isClient, setIsClient] = useState(false)
-  const mediaRecorderRef = useRef<MediaRecorder | null>(null)
-  const timerRef = useRef<NodeJS.Timeout | null>(null)
-  const audioChunksRef = useRef<Blob[]>([])
 
   useEffect(() => {
     setIsClient(true)
   }, [])
-
-  const startRecording = async () => {
-    if (!navigator.mediaDevices || !navigator.mediaDevices.getUserMedia) {
-      alert(
-        'マイク機能がこのブラウザではサポートされていないか、安全でない接続(HTTP)でアクセスしている可能性があります。',
-      )
-      console.error('MediaDevices API not supported.')
-      return
-    }
-    try {
-      const stream = await navigator.mediaDevices.getUserMedia({ audio: true })
-      mediaRecorderRef.current = new MediaRecorder(stream)
-      audioChunksRef.current = []
-
-      mediaRecorderRef.current.ondataavailable = (event) => {
-        audioChunksRef.current.push(event.data)
-      }
-
-      mediaRecorderRef.current.onstop = () => {
-        const audioBlob = new Blob(audioChunksRef.current, { type: 'audio/webm' })
-        setAudioBlob(audioBlob)
-        const audioUrl = URL.createObjectURL(audioBlob)
-        setAudioURL(audioUrl)
-        stream.getTracks().forEach((track) => track.stop()) // Stop the microphone access
-      }
-
-      mediaRecorderRef.current.start()
-      setIsRecording(true)
-      setElapsedTime(0)
-      timerRef.current = setInterval(() => {
-        setElapsedTime((prev) => {
-          if (prev >= 59) {
-            stopRecording()
-            return 60
-          }
-          return prev + 1
-        })
-      }, 1000)
-    } catch (err) {
-      console.error('Error accessing microphone:', err)
-      alert(`マイクへのアクセスでエラーが発生しました: ${err}`)
-    }
-  }
-
-  const stopRecording = () => {
-    if (mediaRecorderRef.current && isRecording) {
-      mediaRecorderRef.current.stop()
-      setIsRecording(false)
-      if (timerRef.current) {
-        clearInterval(timerRef.current)
-      }
-    }
-  }
-
-  const handleToggleRecording = () => {
-    if (isRecording) {
-      stopRecording()
-    } else {
-      setAudioURL(null)
-      setAudioBlob(null)
-      startRecording()
-    }
-  }
 
   const handleSend = async () => {
     if (!audioBlob) return
@@ -101,14 +34,6 @@ export function Llm() {
       alert('音声の送信中にエラーが発生しました。')
     }
   }
-
-  useEffect(() => {
-    return () => {
-      if (timerRef.current) {
-        clearInterval(timerRef.current)
-      }
-    }
-  }, [])
 
   const formatTime = (seconds: number) => {
     const minutes = Math.floor(seconds / 60)
